feat: make it work in jsdoc, rename to completions everywhere
